Added Main Light
Added mainlight object with the global light and the day/night script.
Tested 1 light on a building to turn on when its night time.
Timer for day/night is set to 60 secs, change it in the inspector if you want. Needs at least 10-15 seconds for the animation to work.
