--- Assets/Scripts/Light/DayNightCycle.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 15cf9930d4830f14d86cac46c2ded4f0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
