+++ Assets/Scripts/Light/MainLighting.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6a390741dcbd11f409dfaa95d9ab6cdd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
